+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - CI-508 - Added how dit help sections back into the new nav
 - CI-510 - Removed intro image being required for section on DIT services page
 - CI-518 - Replace how we help on International home page with How DIT help cards section
+- CI-501 - Added international triage contact form from great-domestic-ui
 
 ### Fixed bugs
 - no ticket - Fixed button on Capital invest landing page going to two lines at desktop